xds: move virtual host matcher test to the xdsresource package
